bug fix
Before, taking the weighted average of a vector when all of the weights were zero would result in zero.  This could cause pathological behavior when building a condensed matrix and then repeatedly using it for many generations. Now, instead, the default behavior is to just take the standard arithmetic average.